Move commands to new namespace
