+++ src/AudioBand/Commands/RelayCommand.cs
@@ -1,0 +1,43 @@
+﻿using System;
+using System.Windows.Input;
+
+namespace AudioBand.Commands
+{
+    /// <summary>
+    /// Relay command that can bind to an action and a predicate
+    /// </summary>
+    /// <typeparam name="T">Command parameter</typeparam>
+    /// <inheritdoc cref="ICommand"/>
+    public class RelayCommand<T> : ICommand
+    {
+        private readonly Action<T> _execute;
+        private readonly Predicate<T> _canExecute;
+
+        public RelayCommand(Action<T> execute) : this(execute, null) {}
+
+        public RelayCommand(Action<T> execute, Predicate<T> canExecute)
+        {
+            _execute = execute ?? throw new ArgumentNullException(nameof(execute));
+            _canExecute = canExecute;
+        }
+
+        /// <inheritdoc cref="ICommand.CanExecute"/>
+        public bool CanExecute(object parameter)
+        {
+            return _canExecute?.Invoke((T)parameter) ?? true;
+        }
+
+        /// <inheritdoc cref="ICommand.CanExecuteChanged"/>
+        public event EventHandler CanExecuteChanged
+        {
+            add => CommandManager.RequerySuggested += value;
+            remove => CommandManager.RequerySuggested -= value;
+        }
+
+        /// <inheritdoc cref="ICommand.Execute"/>
+        public void Execute(object parameter)
+        {
+            _execute((T)parameter);
+        }
+    }
+}
