+++ src/AudioBand/Views/Wpf/ResetButton.xaml
@@ -4,6 +4,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:AudioBand.Settings"
+             xmlns:commands="clr-namespace:AudioBand.Commands"
              mc:Ignorable="d" 
              d:DesignHeight="50" d:DesignWidth="50">
     <UserControl.Resources>
@@ -23,7 +24,7 @@
         <Button Grid.Row="0" Grid.Column="0"
                 HorizontalAlignment="Left"
                 Content="Reset" Style="{StaticResource SettingsButton}" 
-                Command="local:SettingsCommands.ResetSetting" CommandParameter="{Binding .}">
+                Command="commands:SettingsCommands.ResetSetting" CommandParameter="{Binding .}">
             <Button.ToolTip>
                 <TextBlock Text="Reset setting to its default values."/>
             </Button.ToolTip>
